Merge branch 'feat/be/code' into 'backend'

See merge request s06-final/S06P31E106!83

+++ backend/src/main/java/com/ssafy/alpaca/api/response/ProblemListRes.java
@@ -19,7 +19,7 @@ public class ProblemListRes {
 
     private String title;
 
-    private Integer level;
+    private Long level;
 
     private LocalDateTime startedAt;
